+++ src/Drivers/geojs.php
@@ -2,6 +2,7 @@
 namespace hamidreza2005\laravelIp\Drivers;
 
 use Illuminate\Support\Collection;
+use Illuminate\Support\Facades\Cache;
 use Illuminate\Support\Facades\Http;
 
 class geojs extends DriverAbstract
@@ -15,10 +16,15 @@ class geojs extends DriverAbstract
      */
     public function setLocation()
     {
+        parent::setLocation();
         $response = Http::get(self::URL.$this->ip().'.json');
         if ($response->failed()){
             throw new \Exception($response);
         }
-        $this->location = collect($response->json());
+        $location = collect($response->json());
+        if (config("ip.caching")){
+            Cache::put("location_".$this->ip(),$location,now()->addWeek());
+        }
+        $this->location = $location;
     }
 }